+++ tools/audit_project/audit_project.py
@@ -19,6 +19,7 @@ Instructions:
 
 """
 
+from copy import deepcopy
 from os import access, R_OK
 from os.path import isfile
 import sys
@@ -28,17 +29,17 @@ from lxml import etree
 
 EXPECTED_OBSERVATIONS = [
     'ob_template', 'ob_0',
-    'ob_1a', 'ob_2b', 'ob_2a', 'ob_2b', 'ob_3a', 'ob_3b', 'ob_4a', 'ob_4b',
+    'ob_1a', 'ob_1b', 'ob_2a', 'ob_2b', 'ob_3a', 'ob_3b', 'ob_4a', 'ob_4b',
     'ob_5a', 'ob_5b', 'ob_6a', 'ob_6b', 'ob_7a', 'ob_7b', 'ob_8a', 'ob_8b',
-    'ob_9a', 'ob_9b', 'ob_10a', 'ob_10b', 'ob_11a', 'ob_11b',
-    'ob_12a', 'ob_12b', 'ob_13a', 'ob_13b', 'ob_14a', 'ob_14b']
+    'ob_9a', 'ob_9b', 'ob_10a', 'ob_10b', 'ob_11a', 'ob_11b', 'ob_12a', 'ob_12b',
+    'ob_13a', 'ob_13b', 'ob_14a', 'ob_14b']
 
 EXPECTED_OBSERVATION_AUTOMATIC_INVITES = [
     # 'ob_template', 'ob_0',   # these two are not automatically sent
-    'ob_1a', 'ob_2b', 'ob_2a', 'ob_2b', 'ob_3a', 'ob_3b', 'ob_4a', 'ob_4b',
+    'ob_1a', 'ob_1b', 'ob_2a', 'ob_2b', 'ob_3a', 'ob_3b', 'ob_4a', 'ob_4b',
     'ob_5a', 'ob_5b', 'ob_6a', 'ob_6b', 'ob_7a', 'ob_7b', 'ob_8a', 'ob_8b',
-    'ob_9a', 'ob_9b', 'ob_10a', 'ob_10b', 'ob_11a', 'ob_11b',
-    'ob_12a', 'ob_12b', 'ob_13a', 'ob_13b', 'ob_14a', 'ob_14b']
+    'ob_9a', 'ob_9b', 'ob_10a', 'ob_10b', 'ob_11a', 'ob_11b', 'ob_12a', 'ob_12b',
+    'ob_13a', 'ob_13b', 'ob_14a', 'ob_14b']
 
 
 # This is the phone number we expect to send our messages 'from'
@@ -50,8 +51,8 @@ STAFF_ALERT_PHONENUM = '61482525929'
 
 
 # These are the pretend contact details we often use for testing
-DEVELOPER_PHONENUMBERS = ['0419153623', '61419153623']
-DEVELOPER_EMAILS = ['danielt@314.net.au', ]
+DEVELOPER_PHONENUMBERS = ['', '']
+DEVELOPER_EMAILS = ['', ]
 
 # We're using pytest for most of the heavy lifting, but we need a way to
 # pass in the project XML file from the cmdline. Pytest makes that difficult
@@ -89,7 +90,7 @@ def proj(request):
 
 def test_twilio_sending_from_australian_mobilenum(proj):
     """Project should be configured to send from our twilo australian number"""
-    pytest.skip("Cannot test project settings from XML yet")
+    pytest.skip("Project settings aren't present in XML file. Need to find alternate test method.")
     # assert == AUSTRALIAN_TWILIO_NUMBER
 
 
@@ -164,12 +165,123 @@ def test_observation_exists(proj, obs):
     assert get_instrument(proj, obs) is not None
 
 
+# TODO rename and test this
+def form_tree_definition(proj, formname):
+    """Turn the form (instrument) definition in the project xml file into
+    a python dictionary tree.
+
+    In the XML structure items are defined in different places and the are
+    various places where 'ref' items have to be looked up to find 'def'.
+
+    Turn it into a nice neat:
+    FormDef
+        ItemGroup(s)
+            Item(s)
+    """
+
+    """
+    Form Def
+    <FormDef OID="Form.ob_template" Name="Ob Template" Repeating="No" redcap:FormName="ob_template">
+        <ItemGroupRef ItemGroupOID="ob_template.timestamp_template" Mandatory="No"/>
+        ...
+    </FormDef>
+
+
+    <ItemGroupDef OID="ob_template.timestamp_template" Name="Ob Template" Repeating="No">
+        <ItemRef ItemOID="timestamp_template" Mandatory="No" redcap:Variable="timestamp_template"/>
+        <ItemRef ItemOID="title_measurements_template" Mandatory="No" redcap:Variable="title_measurements_template"/>
+        ...
+    </ItemGroupDef>
+    <ItemGroupDef OID="ob_template.supqn_tired_template" Name="Confirm and Submit" Repeating="No">
+        ...
+    </ItemGroupDef>
+    ...
+    """
+
+    tree = {}
+    xpath = r"./Study/MetaDataVersion/FormDef/[@OID='%s']" % formname
+    root = proj.find(xpath, proj.nsmap)
+
+    tree = {**tree, **root.attrib}
+    # tree['ItemGroupRef'] = []
+    # tree['ItemGroupDef'] = []
+    tree['ItemGroup'] = []
+
+    for igr in root.getchildren():
+        igr_oid = igr.attrib['ItemGroupOID']
+
+        igd_xpath = r"./Study/MetaDataVersion/ItemGroupDef/[@OID='%s']" % igr_oid
+        igd = proj.find(igd_xpath, proj.nsmap)
+
+        itemgroup = {**igr.attrib, **igd.attrib}
+
+        itemgroup['Item'] = []
+        for itemref in igd.getchildren():
+            item_oid = itemref.attrib['ItemOID']
+
+            itmd_xpath = r"./Study/MetaDataVersion/ItemDef/[@OID='%s']" % item_oid
+            itm = proj.find(itmd_xpath, proj.nsmap)
+
+            item = {
+                **itm.attrib
+            }
+            itemgroup['Item'].append(item)
+
+        tree['ItemGroup'].append(itemgroup)
+
+    return tree
+
+
+# TODO rename and test this
+def rename_tree(form_tree, obcode):
+    """Given a tree template from form_tree_definition('Form.ob_template')
+    recurse downwards renaming keys/vals from ob_template to ob_{code}
+
+    Why? This results in a nice python dictionary object that we can ==
+    compare with pytest and get pretty printed diff.
+    """
+    ob_suffix = obcode.split('_')[1]
+    new_suffix = f'_{ob_suffix}'
+    new = deepcopy(form_tree)
+    new['Name'] = new['Name'].replace('Template', ob_suffix)
+    new['OID'] = new['OID'].replace('_template', new_suffix)
+    new['{https://projectredcap.org}FormName'] = new['{https://projectredcap.org}FormName'].replace('_template', new_suffix)
+
+    for ig in new['ItemGroup']:
+        ig['Name'] = ig['Name'].replace('Template', ob_suffix)
+        ig['OID'] = ig['OID'].replace('_template', new_suffix)
+        ig['ItemGroupOID'] = ig['ItemGroupOID'].replace('_template', new_suffix)
+
+        # ig['{https://projectredcap.org}FormName'] = new['{https://projectredcap.org}FormName'].replace('_template', new_suffix)
+
+        for item in ig['Item']:
+
+            for k, v in item.items():
+                item[k] = v.replace('_template', new_suffix)
+
+            # item['Name'] = new['Name'].replace('_template', new_suffix)
+            # item['{https://projectredcap.org}FormName'] = new['{https://projectredcap.org}FormName'].replace('_template', new_suffix)
+
+    return new
+
+
 @pytest.mark.parametrize("obs", EXPECTED_OBSERVATIONS)
 def test_observation_structure_matches_template(proj, obs):
-    tpl = get_instrument(proj, 'ob_template')
-    obs = get_instrument(proj, obs)
+    # No need to compare template to itself
+    if obs == 'ob_template':
+        return True
 
-    pytest.skip("Not Implemented")
+    # Are the fields the same?
+    # Are the
+    # TODO break out all this code
+
+
+
+    # Should have the same field structure
+    tpl_formdef = form_tree_definition(proj, "Form.ob_template")
+    obs_formdef = form_tree_definition(proj, f"Form.{obs}")
+
+    assert obs_formdef == rename_tree(tpl_formdef, obs)
 
 
 @pytest.mark.parametrize("obs", EXPECTED_OBSERVATIONS)
@@ -179,13 +291,15 @@ def test_observation_has_survey(proj, obs):
 
 
 @pytest.mark.parametrize("obs", EXPECTED_OBSERVATION_AUTOMATIC_INVITES)
-def test_all_observations_have_automated_invite(proj, obs):
+def test_observation_has_automated_invite(proj, obs):
     """Does this observation (eg ob_3a) have an automatic invite defined"""
     assert get_instrument_survey_automated_invite(proj, obs) is not None
 
 
 def test_registration_does_not_have_automated_invite(proj):
-    # Either note defined, or disabled
+    """Registration is triggered by a staff member opening a link to register
+    a new patient. It should not be automatically sent to anyone"""
+    # Either not defined, or disabled (can't work out how to del the old one)
     invite = get_instrument_survey_automated_invite(proj, 'registration')
     assert invite is None or invite.attrib['active'] == '0'
 
@@ -257,13 +371,8 @@ def test_observation_auto_invite_settings(proj, obs):
     """Check a number of specific logic / settings on observation auto invite"""
     invite = get_instrument_survey_automated_invite(proj, obs)
 
-    # Triggers on the correct observation
-    # ob_1a / ob_1b trigger on ob_0
-    # ob_2a triggers on ob_1a
-    # ob_2b triggers on ob_1b
-    # ...and so on
-    expected_trigger_obs = get_obs_preceding_obs(obs)
-    assert invite.attrib['condition_surveycomplete_survey_id'] == expected_trigger_obs
+    # Invites should be triggered on save of ob_0
+    assert invite.attrib['condition_surveycomplete_survey_id'] == 'ob_0'
 
     # The auto invite must be enabled
     assert invite.attrib['active'] == '1'
@@ -282,11 +391,25 @@ def test_observation_auto_invite_settings(proj, obs):
 
     assert invite.attrib['condition_send_next_time'] == expected_time
 
-    # Only sends if they're still under observation
-    # And checks this logic as it's about to be sent
+    # Logic shoudl be enabled (instrument save, 'AND' ....)
+    # Logic should be checked again before send
     assert invite.attrib['condition_andor'] == "AND"
-    assert invite.attrib['condition_logic'] == "[calc_mon_status_observation] = 1"
     assert invite.attrib['reeval_before_send'] == "1"
+
+    # Only sends if they're still under observation
+    active_monitoring_check = '[calc_mon_status_observation] = 1'
+    assert active_monitoring_check in invite.attrib['condition_logic']
+
+    # Confirm the trigger scheduling. This is trick and ugly because of
+    # redcap limitations. Read the design reason for this in
+    # documentation/redcap_design_overview.md  'Observation Invites'
+    # Fragile tests :(
+    send_day = get_obs_day(obs)
+    trigger_day = send_day - 1
+    trigger_logic = f"[calc_mon_status_observation] = 1 and (datediff([mon_admission_date], 'today', 'd') = {trigger_day} or datediff([mon_admission_date], 'today', 'd') = {send_day})"
+
+    assert invite.attrib['condition_logic'] == trigger_logic
+
 
     # Is set to remind the patient
     # Twice, every two hours
@@ -325,6 +448,18 @@ def test_alert_template_exists_combined_patient_alert_sms_exists(proj):
     assert get_alert_template_combined_patient_alert_sms(proj) is not None
 
 
+def test_email_alerts_have_to_address(proj):
+    """All email alerts in the system must have an email-to address.
+    Regression test. One of our tools broke and removed the to address in some
+    existing email alerts
+    """
+    alerts_path = r"./Study/GlobalVariables/redcap:AlertsGroup"
+    alerts = proj.find(alerts_path, proj.nsmap)
+    xpath = f"redcap:Alerts/[@alert_type='EMAIL']"
+    for alert in alerts.findall(xpath, proj.nsmap):
+        assert alert.attrib['email_to'] > ''
+
+
 @pytest.mark.parametrize(", obs", EXPECTED_OBSERVATIONS)
 def test_observation_has_staff_alerts(proj, obs):
 
@@ -354,15 +489,17 @@ def test_observation_staff_alerts_match_template(proj, obs):
     if obs == "ob_template":
         return True
 
+    ignore_fields = ['cron_send_email_on_next_time']
+
     template = get_alert_template_combined_staff_alert_sms(proj)
     alert_title = "Obs Combined Staff alert - %s" % obs
     sms = get_alert(proj, alert_title, form_name=obs, alert_type='SMS')
-    assert obs_alert_matches_template(sms, obs, template)
+    assert obs_alert_matches_template(sms, obs, template, ignore_fields)
 
     template = get_alert_template_combined_staff_alert_email(proj)
     alert_title = "Obs Combined Staff alert email - %s" % obs
     email = get_alert(proj, alert_title, form_name=obs, alert_type='EMAIL')
-    assert obs_alert_matches_template(email, obs, template)
+    assert obs_alert_matches_template(email, obs, template, ignore_fields)
 
 
 @pytest.mark.parametrize(", obs", EXPECTED_OBSERVATIONS)
@@ -375,7 +512,7 @@ def test_observation_patient_alerts_match_template(proj, obs):
     template = get_alert_template_combined_patient_alert_sms(proj)
     alert_title = "Obs Combined Patient alert - %s" % obs
     sms = get_alert(proj, alert_title, form_name=obs, alert_type='SMS')
-    assert obs_alert_matches_template(sms, obs, template)
+    assert obs_alert_matches_template(sms, obs, template, [])
 
 
 @pytest.mark.parametrize(", obs", EXPECTED_OBSERVATIONS)
@@ -389,12 +526,16 @@ def test_late_observation_staff_alert_match_template(proj, obs):
         return True
 
     # These trigger on the previous days alert
+    # Dumb redcap reason. 
+    # See documentation/redcap_design_overview.md  'Late Observation Alert for Staff'
     preceding_obs = get_obs_preceding_obs(obs)
     alert_title = "Late obs staff - %s" % obs
     sms = get_alert(proj, alert_title, form_name=preceding_obs, alert_type='SMS')
-
     template = get_alert_template_late_obs_staff_alert_sms(proj)
-    assert obs_alert_matches_template(sms, obs, template)
+
+    # Ignore the time of day, that will be different for morning/afternoons
+    ignore_fields = ['cron_send_email_on_next_time']
+    assert obs_alert_matches_template(sms, obs, template, ignore_fields)
 
 
 @pytest.mark.parametrize(", obs", EXPECTED_OBSERVATIONS)
@@ -418,11 +559,50 @@ def test_late_observation_staff_alert_triggers_at_5_hours(proj, obs):
         assert False  # unknown obcode
 
 
+def test_registered_patients_should_immediately_be_directed_to_consent_form(proj):
+    """After a patient is registered we then ask them to fill out the Consent
+    instrument"""
+    # This is implemented as an alert that is sent on Registration save
+    # that checks the monitoring group of the new patient
+    # Remember ineligible patients have no need to go to the consent form
+    # but they still get a Registration instrument.
+    TITLE = 'Patient Registration - Consent'
+    sms = get_alert(proj, title=TITLE, form_name='registration', alert_type='SMS')
+    assert sms is not None
+
+    # TODO change this to be an AST comparison so it's less flakey
+    alert_logic = sms.attrib['alert_condition']
+    assert alert_logic == "[mon_group] <> '0'"
+
+    # Should be sent immediately
+    assert sms.attrib['cron_send_email_on'] == 'now'
+
+
+def test_after_completing_consent_patient_should_be_linked_to_first_observation(proj):
+    """After patient completes consent form (and agrees) they should be
+    get a link telling them what to do next (based on monitoring group)
+
+    BIDAILY - should get first observation link
+    WEEKLY - just a thank you and a note to speak to staff"""
+    TITLE = 'Patient Registration - BIDAILY'
+
+    # Should trigger off of consent
+    trigger_instrument = 'consent'
+    sms = get_alert(proj, title=TITLE, form_name=trigger_instrument, alert_type='SMS')
+    assert sms is not None
+
+    # TODO change this to be an AST comparison so it's less flakey
+    # Check that they've agreed, and they're under observation
+    alert_logic = sms.attrib['alert_condition']
+    assert 'cons_agree' in alert_logic
+    assert 'calc_mon_status_observation' in alert_logic
+
+
 def test_staff_are_reminded_to_call_bidaily_patients_after_n_days(proj):
     """We should prompt staff after N days to call our daily monitoring patients"""
 
     TITLE = 'Staff Reminder Patient Call BIDAILY Day 7'
-    sms = get_alert(proj, alert_title=TITLE, form_name=None, alert_type='SMS')
+    sms = get_alert(proj, title=TITLE, form_name='registration', alert_type='SMS')
     assert sms is not None
 
     AFTER_DAYS = 7
@@ -438,7 +618,8 @@ def test_post_discharge_patient_followup_alert_is_scheduled(proj):
     survey"""
     TITLE = 'Post Discharge Patient Follow-up - SMS'
 
-    sms = get_alert(proj, alert_title=TITLE, form_name=None, alert_type='SMS')
+    # TODO change this to be based on Discharge form
+    sms = get_alert(proj, title=TITLE, form_name='discharge', alert_type='SMS')
     assert sms is not None
 
     AFTER_DAYS = 2
@@ -448,29 +629,68 @@ def test_post_discharge_patient_followup_alert_is_scheduled(proj):
     assert sms.attrib['cron_send_email_on_time_lag_minutes'] == "0"
 
 
-def obs_alert_matches_template(alert, obsid, templatealert):
+def test_discharge_reminder_to_staff_should_include_discharge_link(proj):
+    """We email staff telling them to discharge a patient when the time comes.
+    That email should include a link to jump straight to the discharge form
+    in redcap"""
+    TITLE = 'Monitoring Discharge Required staff'
+    sms = get_alert(proj, title=TITLE, form_name="", alert_type='EMAIL')
+    assert sms is not None
+
+    assert '[form-url:discharge]' in sms.attrib['alert_message']
+
+
+def test_patients_awaiting_discharge_should_show_in_awaiting_discharge_report(proj):
+    """We have a report 'Patients awaiting discharge' that should show staff
+    any patients that they need to discharge."""
+
+    TITLE = 'Patients awaiting discharge'
+    rpt = get_report(proj, title=TITLE)
+    assert rpt is not None
+
+    # It probably should match the same patients as the staff dischage reminder
+    # TODO this is a bad test
+    TITLE = 'Monitoring Discharge Required staff'
+    alert = get_alert(proj, title=TITLE, form_name="", alert_type='EMAIL')
+    assert alert is not None
+
+    report_logic = rpt.attrib['advanced_logic']
+    alert_logic = alert.attrib['alert_condition']
+
+    # TODO change this to be an AST comparison so it's less flakey
+    assert report_logic == alert_logic
+
+
+def obs_alert_matches_template(alert, obsid, templatealert, allowed_field_differences):
     """Compare a redcap observation <redcap:Alerts nodes to see if its content
     is equivalent to the _template version of the alert. We don't expect them
     to be identical. The variables names (eg hr_template hr_1a) will differ
     and some meta data like alert IDs and timestamps."""
 
     # These fields we expect to be different
-    IGNORE_FIELDS = ['alert_number', 'form_name', 'alert_condition',
-                     'alert_message', 'alert_title', 'email_subject',
-                     'email_timestamp_sent', 'email_sent']
+    ALWAYS_IGNORE_FIELDS = ['alert_number', 'form_name', 'alert_condition',
+                            'alert_message', 'alert_title', 'email_subject',
+                            'email_timestamp_sent', 'email_sent']
+
+    ignore_fields = ALWAYS_IGNORE_FIELDS + allowed_field_differences
 
     # Compare all the simple fields that should be identical
     alert_vars = {
-        k: v for k, v in alert.attrib.items() if k not in IGNORE_FIELDS}
+        k: v for k, v in alert.attrib.items() if k not in ignore_fields}
 
     template_vars = {
-        k: v for k, v in templatealert.attrib.items() if k not in IGNORE_FIELDS}
+        k: v for k, v in templatealert.attrib.items() if k not in ignore_fields}
 
-    for field in [f for f in templatealert.attrib.keys() if f not in IGNORE_FIELDS]:
+    for field in [f for f in templatealert.attrib.keys() if f not in ignore_fields]:
         # if not alert_vars[field] == template_vars[field]:
         #     breakpoint()
 
         #pytest.xfail("Not Implemented")
+        if not alert_vars[field] == template_vars[field]:
+        #    if template_vars[field] == 'covidhmp@mh.org.au':
+        #    breakpoint()
+            pass
+
         assert alert_vars[field] == template_vars[field]
 
     if not alert_vars == template_vars:
@@ -501,8 +721,21 @@ def obs_alert_matches_template(alert, obsid, templatealert):
 
 
 def test_staff_reminder_patient_day_7_alert_exists(proj):
+    # Staff are reminded to call home observation patients 7 days after they
+    # begin their observation.
     #  Staff Reminder Patient Call BIDAILY Day 7 - SMS
-    pytest.skip("Not Implemented")
+
+    TITLE = 'Staff Reminder Patient Call BIDAILY Day 7'
+
+    # TODO change this to be based on Discharge form
+    sms = get_alert(proj, title=TITLE, form_name='registration', alert_type='SMS')
+    assert sms is not None
+
+    AFTER_DAYS = 7
+    assert sms.attrib['cron_send_email_on'] == "time_lag"
+    assert sms.attrib['cron_send_email_on_time_lag_days'] == str(AFTER_DAYS)
+    assert sms.attrib['cron_send_email_on_time_lag_hours'] == "0"
+    assert sms.attrib['cron_send_email_on_time_lag_minutes'] == "0"
 
 
 def get_instrument_variable_names(proj, id):
@@ -586,11 +819,20 @@ def get_alert_template_late_obs_staff_alert_sms(proj):
     return node
 
 
-def get_alert(proj, alert_title, form_name, alert_type):
+def get_alert(proj, title, form_name, alert_type):
+    """Find REDCap Alert definition node in project xml file"""
     alerts_path = r"./Study/GlobalVariables/redcap:AlertsGroup"
     alerts = proj.find(alerts_path, proj.nsmap)
-    xpath = f"redcap:Alerts/[@form_name='{form_name}'][@alert_title='{alert_title}'][@alert_type='{alert_type}']"
+    xpath = f"redcap:Alerts/[@form_name='{form_name}'][@alert_title='{title}'][@alert_type='{alert_type}']"
     return alerts.find(xpath, proj.nsmap)
+
+
+def get_report(proj, title):
+    """Find REDCap definition node in project xml file"""
+    reports_path = r"./Study/GlobalVariables/redcap:ReportsGroup"
+    reports = proj.find(reports_path, proj.nsmap)
+    xpath = f"redcap:Reports/[@title='{title}']"
+    return reports.find(xpath, proj.nsmap)
 
 
 def obs_is_morning(obcode):
@@ -601,6 +843,21 @@ def obs_is_morning(obcode):
 def obs_is_afternoon(obcode):
     """Given an observation code (eg 'ob_1a', 'ob12_b') is this an afternoon obs?"""
     return obcode[-1] == 'b'
+
+
+def get_obs_day(obcode):
+    """Given an observation code get the day number as integer
+    ob_11b -> 11
+    ob_4a  -> 4
+    """
+    assert valid_ob_code(obcode)
+
+    # The rest follow the previous day (number -1)
+    if obcode == 'ob_0':
+        return 0
+
+    ob_day = int(obcode[3:-1])
+    return ob_day
 
 
 def get_obs_preceding_obs(obcode):
@@ -614,13 +871,16 @@ def get_obs_preceding_obs(obcode):
     """
     assert valid_ob_code(obcode)
 
+    # Nothing comes before ob_0
+    if obcode == 'ob_0':
+        assert False
+
     # First day follows the demo observation
     if obcode in ['ob_1a', 'ob_1b']:
         return 'ob_0'
 
     # The rest follow the previous day (number -1)
-    ob_day = obcode[3:-1]
-    prev_day = int(ob_day) - 1
+    prev_day = get_obs_day(obcode) - 1
 
     if obs_is_morning(obcode):
         return 'ob_%sa' % prev_day
@@ -663,20 +923,71 @@ def test_equiv_template_text():
     assert not equiv_template_text("hello [name_3a]", "hello [name_template]", "ob_1a")
 
 
-def test_xml_doesnt_contain_dev_email(projxmlpath):
-    """Daniel's email should not be present in the project
+def test_get_obs_day():
+    assert get_obs_day('ob_0') == 0
+    assert get_obs_day('ob_1a') == 1
+    assert get_obs_day('ob_2b') == 2
+    assert get_obs_day('ob_14a') == 14
+    assert get_obs_day('ob_0') == 0
+
+    with pytest.raises(AssertionError):
+        get_obs_day('')
+
+    with pytest.raises(AssertionError):
+        get_obs_day('ob_5')
+
+    with pytest.raises(AssertionError):
+        get_obs_day('blaa')
+
+
+def test_get_obs_preceding_obs():
+    assert get_obs_preceding_obs('ob_2a') == 'ob_1a'
+    assert get_obs_preceding_obs('ob_2b') == 'ob_1b'
+    assert get_obs_preceding_obs('ob_10b') == 'ob_9b'
+
+    with pytest.raises(AssertionError):
+        assert get_obs_preceding_obs('ob_0')
+
+    with pytest.raises(AssertionError):
+        get_obs_preceding_obs('')
+
+    with pytest.raises(AssertionError):
+        get_obs_preceding_obs('ob_5')
+
+    with pytest.raises(AssertionError):
+        get_obs_preceding_obs('blaa')
+
+
+def test_project_uses_capital_letters_for_covid(projxmlpath):
+    """We should use COVID-19 not covid Covid covid-19"""
+    with open(projxmlpath) as file:
+        assert "covid" not in file.read()
+    with open(projxmlpath) as file:
+        assert "Covid" not in file.read()
+
+
+def test_project_doesnt_contain_dev_email(projxmlpath):
+    """Daniel's email should not be present in the project's xml file
     TODO: Bad test"""
     for email in DEVELOPER_EMAILS:
         with open(projxmlpath) as file:
             assert email not in file.read()
 
 
-def test_xml_doesnt_contain_dev_phone(projxmlpath):
-    """Daniel's email should not be present in the project
+def test_project_doesnt_contain_dev_phone(projxmlpath):
+    """Daniel's email should not be present in the project's xml file
     TODO: Bad test"""
     for phone in DEVELOPER_PHONENUMBERS:
         with open(projxmlpath) as file:
             assert phone not in file.read()
+
+
+def test_project_doesnt_contain_old_staff_email(projxmlpath):
+    """The pre may 2020 staff email address should not be used anywhere
+    in the project xml file"""
+    with open(projxmlpath) as file:
+        assert "covidhmp@mh.org.au" not in file.read()
+
 
 
 if __name__ == '__main__':
@@ -692,5 +1003,5 @@ if __name__ == '__main__':
 
     # Hand over to pytest
     # HACK. Stuff xml path into a pytest variable, defined in conftest.py
-    pytest.main(["audit_project.py", "-v",
+    pytest.main(["audit_project.py", "-vvvx",
                  "--projxmlfile", proj_path])